moved /context to root folder to clear error for no default export of a compnent in the /app directory
