+++ context/current-npc.tsx
